backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
httpx==0.27.2
python-multipart==0.0.20
pillow==11.2.1
